[Rename] 인증서버 프로젝트명 수정
