+++ auth-server/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'auth-server'
+rootProject.name = 'gold-auth'
