+++ auth-server/src/main/java/com/ryuneng/goldauth/GoldAuthApplication.java
@@ -1,13 +1,13 @@
-package wantedgold.authserver;
+package com.ryuneng.goldauth;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AuthServerApplication {
+public class GoldAuthApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AuthServerApplication.class, args);
+		SpringApplication.run(GoldAuthApplication.class, args);
 	}
 
 }
