+++ auth-server/src/main/java/com/ryuneng/goldauth/domain/user/enums/Role.java
@@ -1,4 +1,4 @@
-package wantedgold.authserver.domain.user.enums;
+package com.ryuneng.goldauth.domain.user.enums;
 
 import lombok.Getter;
