+++ auth-server/src/main/java/com/ryuneng/goldauth/domain/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package wantedgold.authserver.domain.user.repository;
+package com.ryuneng.goldauth.domain.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wantedgold.authserver.domain.user.entity.User;
+import com.ryuneng.goldauth.domain.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
