+++ auth-server/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8888
+  port: ${SERVER_PORT}
 
 spring:
   application:
