+++ auth-server/src/test/java/com/ryuneng/goldauth/GoldAuthApplicationTests.java
@@ -1,10 +1,10 @@
-package wantedgold.authserver;
+package com.ryuneng.goldauth;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AuthServerApplicationTests {
+class GoldAuthApplicationTests {
 
 	@Test
 	void contextLoads() {
